@@ -9,9 +9,10 @@
   <version>1.0-SNAPSHOT</version>
 
   <properties>
-    <java.version>14</java.version>
-    <dropwizard.version>2.1.4</dropwizard.version>
-    <jersey.version>2.37</jersey.version>
+    <java.version>19</java.version>
+    <dropwizard.version>4.0.2</dropwizard.version>
+    <jersey.version>3.0.11</jersey.version>
+    <jakarta.version>2.1.1</jakarta.version>
     <error_prone.version>2.22.0</error_prone.version>
     <mvn-test-plugins.version>3.1.2</mvn-test-plugins.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -34,7 +35,7 @@
 
     <dependency>
       <groupId>org.junit.jupiter</groupId>
-      <artifactId>junit-jupiter-api</artifactId>
+      <artifactId>junit-jupiter</artifactId>
       <version>5.10.0</version>
       <scope>test</scope>
     </dependency>
@@ -79,6 +80,18 @@
         <groupId>org.glassfish.jersey.core</groupId>
         <artifactId>jersey-common</artifactId>
         <version>${jersey.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>jakarta.annotation</groupId>
+        <artifactId>jakarta.annotation-api</artifactId>
+        <version>${jakarta.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>jakarta.activation</groupId>
+        <artifactId>jakarta.activation-api</artifactId>
+        <version>${jakarta.version}</version>
       </dependency>
 
       <dependency>
@@ -134,7 +147,8 @@
         <artifactId>maven-checkstyle-plugin</artifactId>
         <version>3.3.0</version>
         <configuration>
-          <encoding>UTF-8</encoding>
+          <inputEncoding>UTF-8</inputEncoding>
+          <outputEncoding>UTF-8</outputEncoding>
           <consoleOutput>true</consoleOutput>
           <failsOnError>true</failsOnError>
           <violationSeverity>warning</violationSeverity>
@@ -160,7 +174,7 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-enforcer-plugin</artifactId>
-        <version>3.2.1</version>
+        <version>3.4.1</version>
         <executions>
           <execution>
             <id>enforce</id>
