@@ -1,6 +1,7 @@
 package weatherservice.climate;
 
-import javax.validation.constraints.NotNull;
+
+import jakarta.validation.constraints.NotNull;
 
 public class ClimateDto {
 
