@@ -1,21 +1,21 @@
 package weatherservice.daylight;
 
+import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON;
 import static java.time.ZoneOffset.UTC;
 import static java.util.concurrent.TimeUnit.DAYS;
-import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 
 import io.dropwizard.jersey.caching.CacheControl;
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.Response;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.OffsetTime;
 import java.time.ZoneOffset;
 import java.util.Date;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Response;
 
 @Consumes(APPLICATION_JSON)
 @Produces(APPLICATION_JSON)
